move `format_path` to HoudiniLoader Class
