+++ client/ayon_houdini/plugins/load/load_bgeo.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-import os
 import hou
 
 from ayon_houdini.api import (
@@ -39,10 +38,9 @@ class BgeoLoader(plugin.HoudiniLoader):
             file_node.destroy()
 
         # Explicitly create a file node
-        path = self.filepath_from_context(context)
         file_node = container.createNode("file", node_name=node_name)
         file_node.setParms(
-            {"file": self.format_path(path, context["representation"])})
+            {"file": self.format_path(context)})
 
         # Set display on last node
         file_node.setDisplayFlag(True)
@@ -59,18 +57,6 @@ class BgeoLoader(plugin.HoudiniLoader):
             suffix="",
         )
 
-    @staticmethod
-    def format_path(path, representation):
-        """Format file path correctly for single bgeo or bgeo sequence."""
-        # The path is either a single file or sequence in a folder.
-        is_sequence = bool(representation["context"].get("frame"))
-        if is_sequence:
-            path = BgeoLoader.replace_with_frame_token(path)
-
-        path = os.path.normpath(path)
-        path = path.replace("\\", "/")
-        return path
-
     def update(self, container, context):
         repre_entity = context["representation"]
         node = container["node"]
@@ -83,10 +69,7 @@ class BgeoLoader(plugin.HoudiniLoader):
             return
 
         # Update the file path
-        file_path = self.filepath_from_context(context)
-        file_path = self.format_path(file_path, repre_entity)
-
-        file_node.setParms({"file": file_path})
+        file_node.setParms({"file": self.format_path(context)})
 
         # Update attribute
         node.setParms({"representation": repre_entity["id"]})
