+++ client/ayon_houdini/plugins/load/load_image.py
@@ -1,4 +1,3 @@
-import os
 import hou
 
 from ayon_core.pipeline import AYON_CONTAINER_ID
@@ -46,10 +45,6 @@ class ImageLoader(plugin.HoudiniLoader):
     color = "orange"
 
     def load(self, context, name=None, namespace=None, data=None):
-        # Format file name, Houdini only wants forward slashes
-        path = self.filepath_from_context(context)
-        path = self.format_path(path, representation=context["representation"])
-
         # Get the root node
         parent = get_image_ayon_container()
 
@@ -60,7 +55,7 @@ class ImageLoader(plugin.HoudiniLoader):
         node = parent.createNode("file", node_name=node_name)
         node.moveToGoodPosition()
 
-        parms = {"filename1": path}
+        parms = {"filename1": self.format_path(context)}
         parms.update(self.get_colorspace_parms(context["representation"]))
 
         node.setParms(parms)
@@ -85,11 +80,8 @@ class ImageLoader(plugin.HoudiniLoader):
         node = container["node"]
 
         # Update the file path
-        file_path = self.filepath_from_context(context)
-        file_path = self.format_path(file_path, repre_entity)
-
         parms = {
-            "filename1": file_path,
+            "filename1": self.format_path(context),
             "representation": repre_entity["id"],
         }
 
@@ -111,19 +103,6 @@ class ImageLoader(plugin.HoudiniLoader):
 
         if not parent.children():
             parent.destroy()
-
-    @staticmethod
-    def format_path(path, representation):
-        """Format file path correctly for single image or sequence."""
-
-        # The path is either a single file or sequence in a folder.
-        is_sequence = bool(representation["context"].get("frame"))
-        if is_sequence:
-            path = ImageLoader.replace_with_frame_token(path)
-
-        path = os.path.normpath(path)
-        path = path.replace("\\", "/")
-        return path
 
     def get_colorspace_parms(self, representation: dict) -> dict:
         """Return the color space parameters.
