+++ client/ayon_houdini/plugins/load/load_vdb.py
@@ -1,4 +1,3 @@
-import os
 import hou
 
 from ayon_houdini.api import (
@@ -37,9 +36,8 @@ class VdbLoader(plugin.HoudiniLoader):
 
         # Explicitly create a file node
         file_node = container.createNode("file", node_name=node_name)
-        path = self.filepath_from_context(context)
         file_node.setParms(
-            {"file": self.format_path(path, context["representation"])})
+            {"file": self.format_path(context)})
 
         # Set display on last node
         file_node.setDisplayFlag(True)
@@ -56,18 +54,6 @@ class VdbLoader(plugin.HoudiniLoader):
             suffix="",
         )
 
-    @staticmethod
-    def format_path(path, representation):
-        """Format file path correctly for single vdb or vdb sequence."""
-        # The path is either a single file or sequence in a folder.
-        is_sequence = bool(representation["context"].get("frame"))
-        if is_sequence:
-            path = VdbLoader.replace_with_frame_token(path)
-
-        path = os.path.normpath(path)
-        path = path.replace("\\", "/")
-        return path
-
     def update(self, container, context):
         repre_entity = context["representation"]
         node = container["node"]
@@ -80,10 +66,7 @@ class VdbLoader(plugin.HoudiniLoader):
             return
 
         # Update the file path
-        file_path = self.filepath_from_context(context)
-        file_path = self.format_path(file_path, repre_entity)
-
-        file_node.setParms({"file": file_path})
+        file_node.setParms({"file": self.format_path(context)})
 
         # Update attribute
         node.setParms({"representation": repre_entity["id"]})
